renamed section components dirs
