+++ src/components/missionSection/Benefits/BenefitsItem.tsx
