+++ src/components/missionSection/StagesList.tsx
