+++ src/components/missionSection/Title/MissionTitle.tsx
