+++ src/components/missionSection/Title/MobileTitle.tsx
