+++ src/components/ourTeamSection/CategoriesList.tsx
