+++ src/components/ourTeamSection/ListItem.tsx
