+++ src/components/ourTeamSection/Title.tsx
